[js] Update js api.
