+++ pion-sfu-sdk/package.json
@@ -25,7 +25,7 @@
     "centrifuge": "^2.1.3",
     "events": "^3.0.0",
     "protoo-client": "^4.0.3",
-    "random-string": "^0.2.0"
+    "uuid": "^3.3.2"
   },
   "keywords": [
     "pion",
